Added duplicate_key_check_filter class, refactored
duplicate_key_check_filter class is a new filter allowing detection of
duplicate keys in parsed objects. It gives a performance hit, but allows
complex keys and verifies that all levels of complex keys are unique. A
core::error is thrown if not all keys are unique.

Refactoring of parsing code is continuing, where parsers are being
implemented now as core::stream_parser subclasses.

Also, refactoring of output classes was implemented, to hide
implementation functions underneath the `impl` namespace.