Refactored, added MariaDB write support, performance improvements
Refactored calls so a stream_handler's begin() and end() members don't
need to be called when implementing a new parser. They're called in the
conversion operators instead, which means calling convert() on a parser
is not a good idea anymore. Use the << and >> operators instead.

MariaDB write support is added, albeit not with dynamic type analysis.
MySQL types need to be supplied in a (currently) code-documented format
to the writer. MariaDB now supports database-level and table-level
reading, with two parsing classes.