Major refactor and performance improvments
Added switch to abandon C++ STL and use a mostly-compatible subset of
the features. Doing so greatly improved IO performance.

Added core::dump class that operates automatically when using operator<<
with a std::ostream and a core::value as parameters. It prints type,
subtype, and value in a human-readable pretty-printed format.